\name{wdParagraph}
\Rdversion{1.1}
\alias{wdParagraph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Insert a paragraph break.
}
\description{
Insert a paragraph break.
}
\usage{
wdParagraph(wdapp = .R2wd)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wdapp}{
    the handle to Word Application (usually not needed).
}
}
%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{IO}
\keyword{connection}

