\name{wdQuit}
\Rdversion{1.1}
\alias{wdQuit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Close Word and remove the handle.
}
\description{
Close Word and remove the handle.
}
\usage{
wdQuit(wdapp = .R2wd)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wdapp}{
    the handle to the Word Application.
}
}
\details{
  This closes Word. Word may ask back whether you wish to save the graph.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
\dontrun{
wdGet()
wdQuit()
}

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{IO}
\keyword{connection}

