\name{wdSelection}
\Rdversion{1.1}
\alias{wdSelection}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get handle to current Selection in Word Application.
}
\description{
Get handle to current Selection in Word Application.
}
\usage{
wdSelection(wdapp = .R2wd)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wdapp}{
    handle to Word Application (usually not needed)
}
}
\author{
Christian Ritter
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{IO}
\keyword{connection}

